backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy[asyncio]>=2.0.25
asyncpg>=0.29.0
alembic>=1.13.1
pydantic>=2.5.3
pydantic-settings>=2.1.0
email-validator>=2.1.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
httpx>=0.26.0
google-api-python-client>=2.114.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0
